VOL-3419: OpenOLT adapter at scale constantly takes more that 10 seconds to react to flows - Pass information to agent to do the flow replication - Consolidate various locks in the adapter and remove reduntant locks

+++ internal/pkg/olterrors/olterrors.go
@@ -203,7 +203,7 @@ type ErrPersistence struct {
 }
 
 // NewErrPersistence constructs a new error based on the given values
-func NewErrPersistence(operation, entityType string, ID uint32, fields log.Fields, wrapped error) LoggableError {
+func NewErrPersistence(operation, entityType string, ID uint64, fields log.Fields, wrapped error) LoggableError {
 	return &ErrPersistence{
 		ErrAdapter{
 			name: "unable-to-persist",
@@ -265,7 +265,7 @@ type ErrFlowOp struct {
 }
 
 // NewErrFlowOp constructs a new error based on the given values
-func NewErrFlowOp(operation string, ID uint32, fields log.Fields, wrapped error) LoggableError {
+func NewErrFlowOp(operation string, ID uint64, fields log.Fields, wrapped error) LoggableError {
 	return &ErrPersistence{
 		ErrAdapter{
 			name: "unable-to-perform-flow-operation",
